+++ tests/phantomjs/HomepageCept.php
@@ -1,0 +1,5 @@
+<?php 
+$I = new AcceptanceTester($scenario);
+$I->wantTo('access the homepage of the site');
+$I->amOnPage('/');
+$I->seeElement('body.home');