+++ changelog.md
@@ -3,6 +3,8 @@ All notable changes after version 1.6.16 to this project will be documented in t
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ##[unreleased] Unreleased
+### Fixed
+- plugin activation/deactivation in `WPBrowser` module, thanks [Ippey](https://github.com/Ippey) 
 
 ##[1.16.0] 2016-09-05
 ### Added